Remove the parallel tiling code path.
The approach did not work out because it was going wide and then joining for each path, which caused the thread pool to run out of work at the end of each path and go back to sleep. Waking the workers back up was too expensive/had too much latency, causing performance to be underwhelming (not necessarily a regression but the speedup was not worth the use of mutliple threads. For future experiments, a better approch would need to be able to keep the workers fed thoughout at least large portions of the workload to be worth it.